added remaining ecs functionality
the whole thing hews pretty close to austin morlan's implementation (and i copied the component array and component manager types). the system manager is just a simple struct with an update function.

also reimplemented:
- the voronoi tile system, using the "Polygon_Tile" and "Transform" components. the World class now solely generates components.
- moving the view with keys (separate out Input_Accelerator as a base class later)

also fixed that pernicious intermittent crash on entering Game_Play: it was because of the copy constructor, lol. all the private members are static now - consider other solutions.