removed deprecated worldgen/components
- systemgen code in World class
- Entity_Info and Color components

i'm not sold on removing entity_info, but i'm not sure how else to make it comport with the generator as it is right now. try and figure out a better solution... coordinating it through the database is weird, and i don't want to make the return types more complex than they already are. maybe should just go with a POD...