restructured game state
rather than a Game_Play state, there is now Game_Turn and Game_Simulate. this will be specific to the prototype.

in addition, the system manager no longer performs updates. all of that is now handled at the Game level (and in fact, update() is the only override in either state as of now)